+++ src/main/java/com/seailz/playtime/spigot/core/util/PlayTime.java
@@ -1,4 +1,4 @@
-package com.seailz.playtime.core.util;
+package com.seailz.playtime.spigot.core.util;
 
 import org.bukkit.entity.Player;
 
@@ -13,18 +13,18 @@ public class PlayTime {
     }
 
     public JSONUtil getFile() {
-        File file = new File(com.seailz.playtime.PlayTime.getInstance().getDataFolder() + "/data", p.getUniqueId() + ".json");
+        File file = new File(com.seailz.playtime.spigot.PlayTime.getInstance().getDataFolder() + "/data", p.getUniqueId() + ".json");
         return new JSONUtil(file);
     }
 
     public void resetTime() {
-        File file = new File(com.seailz.playtime.PlayTime.getInstance().getDataFolder(), p.getUniqueId() + ".json");
+        File file = new File(com.seailz.playtime.spigot.PlayTime.getInstance().getDataFolder(), p.getUniqueId() + ".json");
         file.delete();
     }
 
     public static void resetGlobalTime() {
         Arrays.stream(
-                new File(com.seailz.playtime.PlayTime.getInstance().getDataFolder() + ".data").listFiles()
+                new File(com.seailz.playtime.spigot.PlayTime.getInstance().getDataFolder() + ".data").listFiles()
         ).forEach(file -> {
             file.delete();
         });
