+++ src/main/resources/config.yml
@@ -1,0 +1,2 @@
+server-identifier: "server"
+token: "hello i'm a bot token"
